Moved flocker_api to example.py
Since we don't do any CLI processing, this is effectively still just
an example of API usage.
