+++ python/example.py
@@ -9,11 +9,13 @@ import tempfile
 class FlockerApi(object):
     DEFAULT_PLUGIN_DIR = os.environ.get('CERT_DIR', '/etc/flocker')
 
-    def __init__(self, api_version = 1):
+    def __init__(self, api_version = 1, debug = False):
         control_service = os.environ.get("CONTROL_SERVICE", "localhost")
         control_port = os.environ.get("CONTROL_PORT", 4523)
 
         self._api_version = api_version
+        self._debug = debug
+        self._last_known_config = None
 
         key_file = os.environ.get("KEY_FILE", "%s/plugin.key" % self.DEFAULT_PLUGIN_DIR)
         cert_file = os.environ.get("CERT_FILE", "%s/plugin.crt" % self.DEFAULT_PLUGIN_DIR)
@@ -65,7 +67,10 @@ class FlockerApi(object):
       if data and not isinstance(data, str):
           data = json.dumps(data).encode('utf-8')
 
-      headers = {"Content-type": "application/json"}
+      headers = { 'Content-type': 'application/json' }
+      if self._last_known_config:
+          headers['X-If-Configuration-Matches'] = self._last_known_config
+
       self._http_client.request(method, endpoint, data,
                                 headers=headers)
 
@@ -74,46 +79,110 @@ class FlockerApi(object):
       status =  response.status
       body =  response.read()
 
+      # Make sure to use the X
+      if 'X-Configuration-Tag' in response.getheaders():
+          self._last_known_config = response.getheaders()['X-Configuration-Tag'].decode('utf-8')
+
       print('Status:', status)
 
-      # If you want debugging
+      # If you want verbose debugging
       # print('Body:', body)
 
       print()
 
-      return json.loads(body.decode('utf-8'))
+      result = json.loads(body.decode('utf-8'))
+
+      if self._debug == True:
+          print(json.dumps(result, sort_keys=True, indent=4))
+
+      return result
+
+    # Specific API requests
+    def get_version(self):
+      version = self.get('version')
+      return version['flocker']
+
+    def create_volume(self, name, size_in_gb, primary_id, profile = None):
+        if not isinstance(size_in_gb, int):
+            print('Error! Size must be an integer!')
+            exit(1)
+
+        data = {
+            'primary': primary_id,
+            'maximum_size': size_in_gb << 30,
+            'metadata': {
+               'name': name
+            }
+        }
+
+        if profile:
+            data['metadata']['clusterhq:flocker:profile'] = profile
+
+        return self.post('configuration/datasets', data)
+
+    def move_volume(self, volume_id, new_primary_id):
+        data = { 'primary': new_primary_id }
+        return self.post('configuration/datasets/%s' % volume_id, data)
+
+    def delete_volume(self, dataset_id):
+        return self.delete('configuration/datasets/%s' % dataset_id)
+
+    def get_volumes(self):
+        return self.get('configuration/datasets')
+
+    def get_nodes(self):
+        return self.get('state/nodes')
+
+    def get_leases(self):
+        return self.get('configuration/leases')
+
+    def release_lease(self, dataset_id):
+        return self.delete('configuration/leases/%s' % dataset_id)
+
+    def acquire_lease(self, dataset_id, node_id, expires = None):
+        data = { 'dataset_id': dataset_id,
+                 'node_uuid': node_id,
+                 'expires': expires }
+        return self.post('configuration/leases', data)
 
 if __name__ == '__main__':
-  api = FlockerApi()
+    api = FlockerApi(debug = True)
 
-  # Show us the version of Flocker
-  version = api.get('version')
-  print("Version:", version['flocker'])
+    # Show us the version of Flocker
+    print("Version:", api.get_version())
 
-  # Get current volumes (datasets)
-  print('Datasets:')
-  datasets = api.get('configuration/datasets')
-  print(json.dumps(datasets, sort_keys=True, indent=4))
+    # Get current volumes (datasets)
+    print('Datasets:')
+    datasets = api.get_volumes()
 
-  # Create a Flocker volume of size 10GB
-  size_in_gb = 10
+    print('Nodes:')
+    nodes = api.get_nodes()
+    first_node = nodes[0]['uuid']
 
-  print('Trying to reuse the primary from returned list')
-  primary_id = datasets[0]['primary']
-  print('Primary:', primary_id)
+    print('Trying to reuse the primary from returned list')
+    primary_id = datasets[0]['primary']
+    print('Primary:', primary_id)
 
-  # XXX: Using shifts to evaluate max bytes
-  #      '<< 30' == '* 2^30' == '* 1024 * 1024 * 1024'
-  data = {
-           'primary': primary_id,
-           'maximum_size': size_in_gb << 30,
-           'metadata': {
-# If your backend supports profiles uncomment the following:
-#             'clusterhq:flocker:profile': 'silver',
-             'name': 'my-test-volume3'
-           }
-         }
+    print('Create volume:')
+    # Create a Flocker volume of size 10GB
+    dataset_create_result = api.create_volume('my-test-volume3',
+                                              10,
+                                              primary_id,
+                                              profile = "gold")
+    volume_id = dataset_create_result['dataset_id']
 
-  print('Create volume:')
-  dataset_create_result = api.post('configuration/datasets', data)
-  print(json.dumps(dataset_create_result, sort_keys=True, indent=4))
+    print('Move volume (to the same node):')
+    api.move_volume(volume_id, primary_id)
+
+    print('Acquire lease:')
+    api.acquire_lease(volume_id, first_node)
+
+    print('Leases:')
+    leases = api.get_leases()
+    lease_id = leases[0]
+
+    print('Release lease:')
+    api.release_lease(volume_id)
+
+    print('Delete volume:')
+    api.delete_volume(volume_id)
